Isolate ezpz MoE
