+++ torchtitan/components/loss.py
@@ -229,12 +229,6 @@ class ChunkedCELoss(BaseLoss):
     class Config(BaseLoss.Config):
         num_chunks: int = 8
         """Number of chunks to split the sequence into."""
-        empty_cache_between_chunks: bool = False
-        """Release cached device memory between chunks for tight-memory runs."""
-        keep_lm_head_unsharded_between_chunks: bool = True
-        """Keep FSDP lm_head unsharded across chunks to avoid repeated all-gathers."""
-        sync_replicated_lm_head_grad: bool = False
-        """All-reduce lm_head grads when lm_head is intentionally not FSDP-wrapped."""
 
     def __init__(
         self,
@@ -245,11 +239,6 @@ class ChunkedCELoss(BaseLoss):
         self.fn: LossFunction = cross_entropy_loss
         self._maybe_compile(compile_config)
         self.num_chunks = config.num_chunks
-        self.empty_cache_between_chunks = config.empty_cache_between_chunks
-        self.keep_lm_head_unsharded_between_chunks = (
-            config.keep_lm_head_unsharded_between_chunks
-        )
-        self.sync_replicated_lm_head_grad = config.sync_replicated_lm_head_grad
         self.lm_head: nn.Module | None = None
 
     def set_lm_head(self, lm_head: nn.Module) -> None:
@@ -317,73 +306,23 @@ class ChunkedCELoss(BaseLoss):
 
         total_loss = hidden_states.new_zeros((), dtype=torch.float32)
 
-        # In the fast mode, keep lm_head unsharded across chunks and coalesce
-        # per-chunk grad sync into a single reduce-scatter at the last chunk.
-        #
-        # In the low-memory mode, explicitly keep FSDP's gradient sync enabled
-        # and reshard around each chunk. Otherwise FSDP accumulates a full
-        # unsharded lm_head gradient, which is too large for 10B_2B_sdpa's
-        # vocab x hidden projection on XPU.
-        keep_lm_head_unsharded = (
-            fsdp_enabled and self.keep_lm_head_unsharded_between_chunks
-        )
-        if keep_lm_head_unsharded:
+        # Disable FSDP reshard on lm_head to keep weight unsharded across
+        # all chunks, avoiding repeated all-gathers. Coalesce per-chunk
+        # grad sync into a single reduce-scatter at the last chunk by
+        # disabling gradient sync for chunks 0..N-2.
+        if fsdp_enabled:
             lm_head.set_reshard_after_forward(False)
             lm_head.set_reshard_after_backward(False)
             lm_head.set_requires_gradient_sync(False, recurse=False)
-        elif fsdp_enabled:
-            lm_head.set_reshard_after_forward(True)
-            lm_head.set_reshard_after_backward(True)
-            lm_head.set_requires_gradient_sync(True, recurse=False)
 
         last_idx = len(h_chunks) - 1
         for i, (h_chunk, label_chunk) in enumerate(zip(h_chunks, label_chunks)):
-            if keep_lm_head_unsharded and i == last_idx:
+            if fsdp_enabled and i == last_idx:
                 lm_head.set_requires_gradient_sync(  # pyrefly: ignore[not-callable]
                     True, recurse=False
                 )
 
-            h_chunk_for_lm_head = h_chunk
-            if fsdp_enabled and not isinstance(h_chunk_for_lm_head, DTensor):
-                lm_head_mesh = None
-                try:
-                    lm_head_state = lm_head._get_fsdp_state()
-                    lm_head_param_group = lm_head_state._fsdp_param_group
-                    if lm_head_param_group is not None:
-                        lm_head_mesh = lm_head_param_group.mesh_info.mesh
-                except Exception:
-                    lm_head_mesh = None
-
-                lm_head_param = getattr(lm_head, "weight", None)
-                if not isinstance(lm_head_param, DTensor):
-                    try:
-                        lm_head_param = next(lm_head.parameters())
-                    except StopIteration:
-                        lm_head_param = None
-                if isinstance(lm_head_param, DTensor):
-                    lm_head_mesh = lm_head_param.device_mesh
-                    placements = tuple(Replicate() for _ in lm_head_param.placements)
-                elif lm_head_mesh is not None:
-                    placements = tuple(Replicate() for _ in range(lm_head_mesh.ndim))
-                else:
-                    placements = None
-
-                if lm_head_mesh is not None and placements is not None:
-                    h_chunk_for_lm_head = DTensor.from_local(
-                        h_chunk,
-                        lm_head_mesh,
-                        placements,
-                        run_check=False,
-                    )
-
-            logits = lm_head(h_chunk_for_lm_head)
-            if isinstance(logits, DTensor) and not isinstance(label_chunk, DTensor):
-                label_chunk = DTensor.from_local(
-                    label_chunk,
-                    logits.device_mesh,
-                    tuple(Replicate() for _ in logits.placements),
-                    run_check=False,
-                )
+            logits = lm_head(h_chunk)
 
             chunk_loss = self.fn(logits, label_chunk)
             if global_valid_tokens is not None:
@@ -392,33 +331,15 @@ class ChunkedCELoss(BaseLoss):
 
             if requires_grad:
                 chunk_loss.backward()
-                h_chunk_grad = h_chunk.grad
-                if h_chunk_grad is None and isinstance(h_chunk_for_lm_head, DTensor):
-                    h_chunk_grad = h_chunk_for_lm_head.grad
-                assert h_chunk_grad is not None
-                grad_accumulator.add(h_chunk_grad)
+                assert h_chunk.grad is not None
+                grad_accumulator.add(h_chunk.grad)
                 h_chunk.grad = None
-                if isinstance(h_chunk_for_lm_head, DTensor):
-                    h_chunk_for_lm_head.grad = None
 
-            if self.empty_cache_between_chunks:
-                del logits, chunk_loss
-                if hidden_states.device.type == "xpu":
-                    torch.xpu.empty_cache()
-                elif hidden_states.device.type == "cuda":
-                    torch.cuda.empty_cache()
-
-        if keep_lm_head_unsharded:
+        if fsdp_enabled:
             lm_head.set_reshard_after_forward(True)
             lm_head.set_reshard_after_backward(True)
             lm_head.set_requires_gradient_sync(True, recurse=False)
             lm_head.reshard()
-        elif fsdp_enabled:
-            lm_head.set_reshard_after_forward(True)
-            lm_head.set_reshard_after_backward(True)
-            lm_head.set_requires_gradient_sync(True, recurse=False)
-        elif self.sync_replicated_lm_head_grad and requires_grad:
-            self._sync_lm_head_grad(lm_head)
         if not requires_grad:
             return total_loss
 
@@ -446,21 +367,6 @@ class ChunkedCELoss(BaseLoss):
         return _DecoderOutputGradientBackProp.apply(
             hidden_states, accumulated_grad, total_loss
         )
-
-    @staticmethod
-    def _sync_lm_head_grad(lm_head: nn.Module) -> None:
-        """Synchronize replicated lm_head gradients across the default group."""
-        from torch.distributed.tensor import DTensor
-
-        if (
-            not torch.distributed.is_available()
-            or not torch.distributed.is_initialized()
-        ):
-            return
-        for param in lm_head.parameters():
-            if param.grad is None or isinstance(param.grad, DTensor):
-                continue
-            torch.distributed.all_reduce(param.grad, op=torch.distributed.ReduceOp.SUM)
 
 
 class _DecoderOutputGradientBackProp(torch.autograd.Function):
