+++ torchtitan/experiments/ezpz/moe/__init__.py
@@ -4,7 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-import dataclasses
 from collections.abc import Callable
 from functools import partial
 from typing import Literal
@@ -17,20 +16,75 @@ from torchtitan.experiments.ezpz.agpt import (
     _ezpz_get_attention_config,
 )
 from torchtitan.models.common import Embedding, Linear, RMSNorm, RoPE, TransformerBlock
-from torchtitan.models.common.config_utils import (
-    make_experts_config,
-    make_ffn_config,
-    make_moe_config,
-    make_router_config,
-)
+from torchtitan.models.common.config_utils import make_ffn_config
 from torchtitan.models.common.param_init import depth_scaled_std
 from torchtitan.protocols.model_spec import ModelSpec
 
 from .experts import ExpertComputeBackend, EzpzGroupedExperts
 from .model import Attention, moeModel, moeTransformerBlock
+from .moe import MoE, TokenChoiceTopKRouter
+from .token_dispatcher import AllToAllTokenDispatcher, DeepEPTokenDispatcher
 
 from .parallelize import parallelize_moe
 from .state_dict_adapter import moeStateDictAdapter
+
+
+def make_ezpz_router_config(
+    *,
+    dim: int,
+    num_experts: int,
+    gate_param_init: dict[str, Callable],
+    top_k: int = 1,
+    score_func: Literal["sigmoid", "softmax"] = "sigmoid",
+    route_norm: bool = False,
+    route_scale: float = 1.0,
+    num_expert_groups: int | None = None,
+    num_limited_groups: int | None = None,
+    bias: bool = False,
+) -> TokenChoiceTopKRouter.Config:
+    return TokenChoiceTopKRouter.Config(
+        num_experts=num_experts,
+        gate=Linear.Config(
+            in_features=dim,
+            out_features=num_experts,
+            bias=bias,
+            param_init=gate_param_init,
+        ),
+        top_k=top_k,
+        score_func=score_func,
+        route_norm=route_norm,
+        route_scale=route_scale,
+        num_expert_groups=num_expert_groups,
+        num_limited_groups=num_limited_groups,
+    )
+
+
+def make_ezpz_token_dispatcher_config(
+    *,
+    num_experts: int,
+    top_k: int,
+    score_before_experts: bool = True,
+    comm_backend: str,
+    non_blocking_capacity_factor: float | None = None,
+):
+    if comm_backend in ("deepep", "hybridep"):
+        return DeepEPTokenDispatcher.Config(
+            num_experts=num_experts,
+            top_k=top_k,
+            score_before_experts=score_before_experts,
+            comm_backend=comm_backend,
+            non_blocking_capacity_factor=non_blocking_capacity_factor,
+        )
+    if comm_backend == "standard":
+        return AllToAllTokenDispatcher.Config(
+            num_experts=num_experts,
+            top_k=top_k,
+            score_before_experts=score_before_experts,
+        )
+    raise ValueError(
+        f"Unknown comm_backend: {comm_backend!r}. "
+        "Must be one of 'standard', 'deepep', 'hybridep'."
+    )
 
 
 def make_ezpz_experts_config(
@@ -48,24 +102,36 @@ def make_ezpz_experts_config(
     """Build an EzpzGroupedExperts.Config from the same args as upstream
     `make_experts_config`, plus a `compute_backend` selector.
     """
-    base = make_experts_config(
+    return EzpzGroupedExperts.Config(
         dim=dim,
         hidden_dim=hidden_dim,
         num_experts=num_experts,
-        top_k=top_k,
         param_init=param_init,
-        score_before_experts=score_before_experts,
-        comm_backend=comm_backend,
-        non_blocking_capacity_factor=non_blocking_capacity_factor,
-    )
-    # Re-wrap as the ezpz subclass Config so the runtime build instantiates
-    # EzpzGroupedExperts (which understands `compute_backend`).
-    field_values = {
-        f.name: getattr(base, f.name) for f in dataclasses.fields(base) if f.init
-    }
-    return EzpzGroupedExperts.Config(
-        **field_values,
+        token_dispatcher=make_ezpz_token_dispatcher_config(
+            num_experts=num_experts,
+            top_k=top_k,
+            score_before_experts=score_before_experts,
+            comm_backend=comm_backend,
+            non_blocking_capacity_factor=non_blocking_capacity_factor,
+        ),
         compute_backend=compute_backend,
+    )
+
+
+def make_ezpz_moe_config(
+    *,
+    num_experts: int = 8,
+    router: TokenChoiceTopKRouter.Config,
+    experts: EzpzGroupedExperts.Config,
+    shared_experts=None,
+    load_balance_coeff: float | None = 1e-3,
+) -> MoE.Config:
+    return MoE.Config(
+        num_experts=num_experts,
+        load_balance_coeff=load_balance_coeff,
+        router=router,
+        experts=experts,
+        shared_experts=shared_experts,
     )
 
 
@@ -248,9 +314,9 @@ def _build_moe_layers(
             moe_cfg = None
         else:
             ffn_cfg = None
-            moe_cfg = make_moe_config(
+            moe_cfg = make_ezpz_moe_config(
                 num_experts=num_experts,
-                router=make_router_config(
+                router=make_ezpz_router_config(
                     dim=dim,
                     num_experts=num_experts,
                     gate_param_init=_depth_init(layer_id),
@@ -999,7 +1065,6 @@ def model_registry(
 ) -> ModelSpec:
     from torchtitan.components.quantization import QuantizationConverter
     from torchtitan.distributed.pipeline_parallel import pipeline_llm
-    from torchtitan.models.common.config_utils import make_token_dispatcher_config
 
     config = moe_configs[flavor]()
 
@@ -1009,7 +1074,7 @@ def model_registry(
     for layer_cfg in config.layers:
         if layer_cfg.moe is not None:
             experts_cfg = layer_cfg.moe.experts
-            experts_cfg.token_dispatcher = make_token_dispatcher_config(
+            experts_cfg.token_dispatcher = make_ezpz_token_dispatcher_config(
                 num_experts=experts_cfg.num_experts,
                 top_k=experts_cfg.token_dispatcher.top_k,
                 score_before_experts=experts_cfg.token_dispatcher.score_before_experts,
